# perf: avoid redundant slice allocation in leafHash assignment